+++ src/renderer/components/content-types/workspace/omnibox/OmniboxWorkspace.css
